Move timestamp conversion to its own test
